+++ tachyon/math/finite_fields/quadratic_extension_field.h
@@ -76,7 +76,7 @@ class QuadraticExtensionField
 
   constexpr bool IsMinusOne() const { return c0_.IsMinusOne() && c1_.IsZero(); }
 
-  constexpr static uint64_t ExtensionDegree() {
+  constexpr static uint32_t ExtensionDegree() {
     return 2 * BaseField::ExtensionDegree();
   }
 
@@ -100,7 +100,7 @@ class QuadraticExtensionField
     return c0_.Square() - Config::MulByNonResidue(c1_.Square());
   }
 
-  constexpr Derived& FrobeniusMapInPlace(uint64_t exponent) {
+  constexpr Derived& FrobeniusMapInPlace(uint32_t exponent) {
     c0_.FrobeniusMapInPlace(exponent);
     c1_.FrobeniusMapInPlace(exponent);
     c1_ *=
