+++ tachyon/math/finite_fields/quartic_extension_field.h
@@ -102,7 +102,7 @@ class QuarticExtensionField : public CyclotomicMultiplicativeSubgroup<Derived> {
     return c0_.IsMinusOne() && c1_.IsZero() && c2_.IsZero() && c3_.IsZero();
   }
 
-  constexpr static uint64_t ExtensionDegree() {
+  constexpr static uint32_t ExtensionDegree() {
     return 4 * BaseField::ExtensionDegree();
   }
 
@@ -131,7 +131,7 @@ class QuarticExtensionField : public CyclotomicMultiplicativeSubgroup<Derived> {
     return self_to_p.c0();
   }
 
-  constexpr Derived& FrobeniusMapInPlace(uint64_t exponent) {
+  constexpr Derived& FrobeniusMapInPlace(uint32_t exponent) {
     c0_.FrobeniusMapInPlace(exponent);
     c1_.FrobeniusMapInPlace(exponent);
     c2_.FrobeniusMapInPlace(exponent);
